refactor(xrpl): move `XRPLClient` into its own module
